Rename css dir to scss
